Arreglo de superposición de mesas.

+++ database/migrations/2023_01_11_125056_create_mesas_table.php
@@ -16,8 +16,8 @@ return new class extends Migration {
             $table->id();
             $table->timestamps();
             $table->string('estado', 50);
-            $table->integer('x');
-            $table->integer('y');
+            $table->float('x');
+            $table->float('y');
         });
     }
